requirements.txt
beautifulsoup4==4.12.2
Requests==2.31.0
python-dotenv~=1.0.0